refactor(nbd-cow): split device lifecycle internals
